HQT on VLLM - move matmul and softmax to hpu utils and revert logits use of matmul class
